+++ ttnn/cpp/ttnn/operations/data_movement/CMakeLists.txt
@@ -27,7 +27,7 @@ file(
     pad/device/kernels/*
     permute/device/kernels/*
     reshape_on_device/device/kernels/*
-    repeat/device/device/*
+    repeat/device/kernels/repeat_*_rm.cpp
     reshape_view/device/device/*
     scatter/device/kernels/*
     sort/device/kernels/*
@@ -150,7 +150,7 @@ target_sources(
         permute/device/permute_rm_program_factory.cpp
         permute/device/permute_tiled_program_factory.cpp
         permute/permute.cpp
-        repeat/device/host/repeat_program_factory.cpp
+        repeat/device/repeat_program_factory.cpp
         repeat/device/repeat_device_operation.cpp
         repeat/repeat.cpp
         repeat_interleave/repeat_interleave.cpp
