+++ ttnn/cpp/ttnn/operations/data_movement/repeat/device/repeat_device_operation.hpp
@@ -6,8 +6,8 @@
 
 #include "ttnn/run_operation.hpp"
 #include "ttnn/decorators.hpp"
-#include "repeat_operation_types.hpp"
-#include "host/repeat_program_factory.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_device_operation_types.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_program_factory.hpp"
 
 namespace ttnn::operations::data_movement::repeat {
