+++ ttnn/cpp/ttnn/operations/data_movement/repeat/device/repeat_program_factory.cpp
@@ -18,7 +18,7 @@
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/types.hpp"
 
-#include "repeat_program_factory.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_program_factory.hpp"
 
 constexpr uint32_t READ_ALIGNMENT = 64;
 
@@ -84,7 +84,7 @@ RepeatProgramFactorySecondDim::cached_program_t RepeatProgramFactorySecondDim::c
 
     tt::tt_metal::KernelHandle reader_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/data_movement/repeat/device/device/repeat_last_dim_rm.cpp",
+        "ttnn/cpp/ttnn/operations/data_movement/repeat/device/kernels/repeat_last_dim_rm.cpp",
         total_cores,
         tt::tt_metal::ReaderDataMovementConfig(compile_time_args));
     uint32_t done = 0;
@@ -174,7 +174,7 @@ RepeatProgramFactoryLastDim::cached_program_t RepeatProgramFactoryLastDim::creat
 
     tt::tt_metal::KernelHandle reader_kernel_id = tt::tt_metal::CreateKernel(
         program,
-        "ttnn/cpp/ttnn/operations/data_movement/repeat/device/device/repeat_higher_dim_rm.cpp",
+        "ttnn/cpp/ttnn/operations/data_movement/repeat/device/kernels/repeat_higher_dim_rm.cpp",
         total_cores,
         tt::tt_metal::ReaderDataMovementConfig(compile_time_args));
     uint32_t done = 0;
