+++ ttnn/cpp/ttnn/operations/data_movement/repeat/device/repeat_program_factory.hpp
@@ -6,7 +6,7 @@
 
 #include <tt-metalium/core_coord.hpp>
 #include "ttnn/device_operation.hpp"
-#include "ttnn/operations/data_movement/repeat/device/repeat_operation_types.hpp"
+#include "ttnn/operations/data_movement/repeat/device/repeat_device_operation_types.hpp"
 
 namespace ttnn::operations::data_movement::repeat::program {
